changed from int to long

+++ app/src/main/java/com/example/csc250_spring2021_mergesort/MainActivity.java
@@ -27,7 +27,7 @@ public class MainActivity extends AppCompatActivity
         this.answerTV.setText(String.valueOf(fact(num)));
 
     }
-        private int fact(int num)
+        private long fact(long num)
         {
 
             if(num <= 1)
